refactor(config_routes): extract skills.py from _factory.py (5/8)
Schritt 5 von 8 im Split-Plan
(`docs/superpowers/plans/2026-04-29-config-routes-split.md`).

What moved
- `_register_skill_routes()` (~365 LOC) -> `cognithor.channels.config_routes.skills`. Marketplace, Updater, Commands, Skill-CLI, Connectors, Workflows, Models, i18n, Setup-Wizard.
- `_register_skill_registry_routes()` (~250 LOC) -> selbe Datei. `/api/v1/skill-registry/*` Endpoints.
- `_register_hermes_routes()` (~70 LOC) -> selbe Datei. agentskills.io / Hermes-Kompatibilitaets-Layer.

`_factory.py` schrumpft von 3 603 -> 2 917 Zeilen (-19 %).

Per-Sub-Module-Imports (`skills.py`)
- Stdlib: `contextlib`, `pathlib.Path`.
- Try-Imports: `Request`, `HTTPException`.
- Typing: `Any`. F401-clean.

`_factory.py` Cleanup
- `from pathlib import Path` (Top-Level) -> `if TYPE_CHECKING: from pathlib import Path`. `Path` wird nur als Annotation in `_load_yaml`/`_save_yaml` Closures gebraucht; mit `from __future__ import annotations` reicht die TC-Variante.

Aufruf-Reihenfolge in `create_config_routes()` exakt wie zuvor.

Verification
- `pytest test_route_inventory_unchanged` -> passed (303/303 endpoints).
- Konsumenten-Sweep ueber 11 Test-Files -> 396 passed.
- `ruff check` -> All checks passed.
- `ruff format --check` -> 9 files already formatted.
